Refactors for readibility and centralization.
