+++ docroot/modules/custom/va_gov_workflow_assignments/src/Plugin/Block/EntityMetaDisplay.php
@@ -13,9 +13,8 @@ use Drupal\Core\Url;
 use Drupal\node\NodeInterface;
 use Drupal\taxonomy\TermInterface;
 use Drupal\va_gov_backend\Service\ExclusionTypesInterface;
-// use Drupal\va_gov_backend\Service\VaGovUrl;
 use Drupal\va_gov_backend\Service\VaGovUrlInterface;
-use Drupal\va_gov_lovell\EventSubscriber\LovellEventSubscriber;
+use Drupal\va_gov_lovell\LovellOps;
 use Drupal\va_gov_workflow_assignments\Service\EditorialWorkflowContentRepository;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
@@ -101,8 +100,6 @@ class EntityMetaDisplay extends BlockBase implements ContainerFactoryPluginInter
     $this->renderer = $renderer;
   }
 
-
-
   /**
    * {@inheritDoc}
    */
@@ -118,49 +115,6 @@ class EntityMetaDisplay extends BlockBase implements ContainerFactoryPluginInter
       $container->get('va_gov_workflow_assignments.editorial_workflow'),
       $container->get('renderer')
     );
-  }
-
-  /**
-   * Returns the type of special case or null
-   *
-   * @param Drupal\node\NodeInterface
-   *   A node.
-   *
-   * @return string
-   *   Type of special case (or empty).
-   */
-  private function isLovell (string $section_id): bool {
-    $is_lovell = false;
-    $lovell_sections = LovellEventSubscriber::LOVELL_SECTIONS;
-    if (array_key_exists($section_id, $lovell_sections)) {
-      $is_lovell = true;
-    }
-    return $is_lovell;
-  }
-  private function buildLovellUrlWithCorrectPrefix(NodeInterface $node, string $prefix): string {
-    $url = "";
-    $original_path = $node->toUrl()->toString();
-    $pattern_to_trim = "/^\/([a-z]|\-)*/i";
-    $trimmed_path = preg_replace($pattern_to_trim,"/" . $prefix, $original_path);
-    $va_gov_url_front_end_url = $this->vaGovUrl->getVaGovFrontEndUrl();
-    $url = $va_gov_url_front_end_url . $trimmed_path;
-    return $url;
-  }
-  private function getPathsToPrint(NodeInterface $node) {
-    $paths_to_print = [];
-    $section_id = $node->get('field_administration')->target_id;
-    $is_lovell = $this->isLovell($section_id);
-    if ($is_lovell) {
-      $valid_prefixes = LovellEventSubscriber::getValidPrefixes($section_id);
-      foreach ($valid_prefixes as $prefix) {
-        $paths_to_print[] = $this->buildLovellUrlWithCorrectPrefix($node, $prefix);
-        }
-    }
-    else {
-      $va_gov_url = $this->vaGovUrl->getVaGovFrontEndUrlForEntity($node);
-      $paths_to_print[] = $va_gov_url;
-    }
-    return $paths_to_print;
   }
 
   /**
@@ -185,8 +139,8 @@ class EntityMetaDisplay extends BlockBase implements ContainerFactoryPluginInter
       $block_items['Owner'] = $this->getSectionHierarchyBreadcrumbLinks($node);
     }
     if ($this->vaGovUrlShouldBeDisplayed($node)) {
-      $paths_to_print = $this->getPathstoPrint($node);
-      foreach ($paths_to_print as $va_gov_url) {
+      $va_gov_urls_to_display = $this->getUrlsToDisplay($node);
+      foreach ($va_gov_urls_to_display as $va_gov_url) {
         if ($this->vaGovUrl->vaGovFrontEndUrlForEntityIsLive($node)) {
           $link = Link::fromTextAndUrl($va_gov_url, Url::fromUri($va_gov_url))->toRenderable();
           $link['#attributes'] = ['class' => 'va-gov-url'];
@@ -211,7 +165,7 @@ class EntityMetaDisplay extends BlockBase implements ContainerFactoryPluginInter
       if ($block_key === 'Owner') {
         $output .= '<div><span class="va-gov-entity-meta__title"><strong>' . $block_key . ': </strong></span><span class="va-gov-entity-meta__content">' . $block_item . '</span></div>';
       }
-      else if ($block_key === 'VA.gov URL') {
+      elseif ($block_key === 'VA.gov URL') {
         foreach ($block_item as $va_gov_url) {
           $output .= $this->t('<div><span class="va-gov-entity-meta__title"><strong>@block_key: </strong></span><span class="va-gov-entity-meta__content">@va_gov_url</span></div>',
           [
@@ -220,15 +174,13 @@ class EntityMetaDisplay extends BlockBase implements ContainerFactoryPluginInter
           ]);
         }
 
-
       }
-      else  {
+      else {
         $output .= $this->t('<div><span class="va-gov-entity-meta__title"><strong>@block_key: </strong></span><span class="va-gov-entity-meta__content">@block_item</span></div>',
         [
           '@block_key' => $block_key,
           '@block_item' => $block_item,
         ]);
-
 
       }
     }
@@ -331,10 +283,34 @@ class EntityMetaDisplay extends BlockBase implements ContainerFactoryPluginInter
     )->toString();
   }
 
-  private function createLovellLink(NodeInterface $node) {
-    return true;
+  /**
+   * Returns the URLs to print (if any).
+   *
+   * @param Drupal\node\NodeInterface $node
+   *   A node.
+   *
+   * @return array
+   *   URL(s) to print.
+   */
+  private function getUrlsToDisplay(NodeInterface $node): array {
+    $va_gov_urls_to_display = [];
+    $section_id = $node->get('field_administration')->target_id;
+    // If this is Lovell, we need to check for
+    // multiple front-end urls and the right prefix.
+    $lovell_type = LovellOps::getLovellType($node);
+    if ($lovell_type !== '') {
+      $valid_prefixes = LovellOps::getValidPrefixes($section_id);
+      foreach ($valid_prefixes as $prefix) {
+        $va_gov_url_front_end_url = $this->vaGovUrl->getVaGovFrontEndUrl();
+        $va_gov_urls_to_display[] = LovellOps::buildLovellUrlWithCorrectPrefix($node, $prefix, $va_gov_url_front_end_url);
+      }
+    }
+    else {
+      $va_gov_url = $this->vaGovUrl->getVaGovFrontEndUrlForEntity($node);
+      $va_gov_urls_to_display[] = $va_gov_url;
+    }
+    return $va_gov_urls_to_display;
   }
-
 
   /**
    * Determine whether the va.gov URL should be displayed.
@@ -364,7 +340,5 @@ class EntityMetaDisplay extends BlockBase implements ContainerFactoryPluginInter
 
     return TRUE;
   }
-
-
 
 }
